requirements.txt
python-telegram-bot==13.11
selenium==4.1.2

pygame==2.1.2
fastapi
uvicorn[standard]
